Change the HTML structure and apply flex-box
Index.HTML did not have a parent element and the flex-box was not applying correctly,
resrtuctured it to have parent and child elements and added more specific class and
ID selectors. Resizing the the window was not keeping the formatting so added wrap
to it.

Styles.css was not layed out clearly so have added each rule to a new line making
it easier to read and edit. Added a workaround to buttons 2 and 3 in the way of top
and bottom margins the allgin them, will look into a better fix later in my studies.